requirements.txt
numpy==1.23.3
opencv-python==4.6.0.66
Pillow==9.2.0
slack-sdk==3.19.1
slack_bolt>=1.6.1
certifi